+++ data_models.py
@@ -35,4 +35,4 @@ class Transaction:
 class SiteStatus:
     status: bool  # True if site is up, False if down
     last_failure_time: int
-    site_log: List[Tuple[float, bool]]
+    site_log: List[Tuple[bool, int]]
