+++ transaction_manager.py
@@ -1,20 +1,14 @@
 from typing import Dict
 
-from data_broker import DataBroker
 from data_models import *
-from site_broker import SiteBroker
 from site_manager import SiteManager
 
 
 class TransactionManager:
-    def __init__(self):
-        self.data_broker = DataBroker()
-        self.site_broker = SiteBroker()
+    def __init__(self, site_manager: SiteManager):
+        self.site_manager = site_manager
         self.transaction_map: Dict[str, Transaction] = {}  # t_id -> transaction information
         self.conflict_graph: Dict[str, Set[str]] = {}  # transaction_id -> set of conflicting transaction_ids
-        self.site_managers: Dict[int, SiteManager] = {
-            i: SiteManager(i) for i in range(1, 11)
-        }
 
     def begin(self, t_id: str, timestamp: int):
         self.transaction_map[t_id] = Transaction(
@@ -32,7 +26,7 @@ class TransactionManager:
             return False
 
         # Get available sites for this data item
-        available_sites = self.site_broker.get_available_sites(data_id, self.data_broker)
+        available_sites = self.site_manager.get_available_sites(data_id)
 
         # TODO: If no sites are available, wait - how do we handle this??
         # TODO: Maybe have a list of pending reads that gets scanned each time a site recovers
@@ -44,7 +38,7 @@ class TransactionManager:
 
         # Try to read from any available site
         for site_id in available_sites:
-            value = self.site_managers[site_id].read(t_id, data_id)
+            value = self.site_manager.get_site(site_id).read(t_id, data_id)
             # TODO: should we update transaction reads regardless of available sites?
             # In that case, we need to move the next few lines of code to the top of self.read()
             if value is not None:
@@ -58,7 +52,7 @@ class TransactionManager:
             return False
 
         # Get available sites for this data item
-        available_sites = self.site_broker.get_available_sites(data_id, self.data_broker)
+        available_sites = self.site_manager.get_available_sites(data_id)
 
         # TODO: what should be done if no sites are available for write ?
         if not available_sites:
@@ -68,7 +62,8 @@ class TransactionManager:
         # Write to all available sites
         success = False
         for site_id in available_sites:
-            if self.site_managers[site_id].write(t_id, data_id, value, timestamp):
+            site = self.site_manager.get_site(site_id)
+            if site.write(t_id, data_id, value, timestamp):
                 success = True
 
         if success:
@@ -127,31 +122,13 @@ class TransactionManager:
         else:
             transaction.status = TransactionStatus.COMMITTED
             # Persist changes to all sites
-            for site_id in self.site_managers.keys():
-                if self.site_broker.is_site_up(site_id):
-                    self.site_managers[site_id].persist(t_id, timestamp)
+            # TODO: Need to change below logic
+            for site_id in range(1, 11):
+                if self.site_manager.is_site_up(site_id):
+                    self.site_manager.get_site(site_id).persist(t_id, timestamp)
             print(f"Transaction {t_id} commits")
 
         # TODO: After the transaction commits, we should remove it from the adjacency list
-
-    def fail(self, site_id: int, timestamp: int):
-        self.site_broker.site_status[site_id].status = False
-        self.site_broker.site_status[site_id].last_failure_time = timestamp
-        self.site_broker.site_status[site_id].site_log.append((False, timestamp))
-        self.site_managers[site_id].fail()
-        print(f"Site {site_id} fails")
-
-    def recover(self, site_id: int, timestamp: int):
-        self.site_broker.site_status[site_id].status = True
-        self.site_broker.site_status[site_id].site_log.append((True, timestamp))
-        self.site_managers[site_id].recover()
-        # TODO: check which pending reads can be completed
-        print(f"Site {site_id} recovers")
-
-    def dump(self):
-        for site_id in self.site_managers.keys():
-            if self.site_broker.is_site_up(site_id):
-                self.site_managers[site_id].dump()
 
     def _is_invalid(self, t_id: str):
         if t_id not in self.transaction_map:
